+++ crates/polars-python/src/dataframe/construction.rs
@@ -1,9 +1,12 @@
 use polars::frame::row::{rows_to_schema_supertypes, rows_to_supertypes, Row};
+use polars::prelude::*;
 use pyo3::prelude::*;
+use pyo3::types::PyDict;
 
-use super::*;
+use super::PyDataFrame;
 use crate::conversion::any_value::py_object_to_any_value;
 use crate::conversion::{vec_extract_wrapped, Wrap};
+use crate::error::PyPolarsErr;
 use crate::interop;
 
 #[pymethods]
