+++ crates/polars-python/src/series/mod.rs
@@ -1,32 +1,28 @@
+#[cfg(feature = "pymethods")]
 mod aggregation;
+#[cfg(feature = "pymethods")]
 mod arithmetic;
+#[cfg(feature = "pymethods")]
 mod buffers;
+#[cfg(feature = "pymethods")]
 mod c_interface;
+#[cfg(feature = "pymethods")]
 mod comparison;
+#[cfg(feature = "pymethods")]
 mod construction;
+#[cfg(feature = "pymethods")]
 mod export;
+#[cfg(feature = "pymethods")]
 mod general;
+#[cfg(feature = "pymethods")]
 mod import;
+#[cfg(feature = "pymethods")]
 mod numpy_ufunc;
+#[cfg(feature = "pymethods")]
 mod scatter;
 
-use std::io::Cursor;
-
-use polars_core::chunked_array::cast::CastOptions;
-use polars_core::series::IsSorted;
-use polars_core::utils::flatten::flatten_series;
-use polars_core::{with_match_physical_numeric_polars_type, with_match_physical_numeric_type};
-use pyo3::exceptions::{PyIndexError, PyRuntimeError, PyValueError};
-use pyo3::prelude::*;
-use pyo3::types::PyBytes;
-use pyo3::Python;
-
-use crate::dataframe::PyDataFrame;
-use crate::error::PyPolarsErr;
-use crate::map::series::{call_lambda_and_extract, ApplyLambda};
-use crate::prelude::*;
-use crate::py_modules::POLARS;
-use crate::raise_err;
+use polars::prelude::Series;
+use pyo3::pyclass;
 
 #[pyclass]
 #[repr(transparent)]
